+++ rollingFood/package.json
@@ -9,6 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "animate.css": "^4.1.1",
     "antd": "^4.22.4",
     "axios": "^0.27.2",
     "bootstrap": "^5.2.0",
