--- rollingFood/src/components/card.css
@@ -1,53 +1,0 @@
-.card:hover {
-    box-shadow: 5px 10px 20px 1px rgba(255, 255, 255, 0.273) !important;
-    transition: all 0.7s linear;
-  }
-  
-  .card-body {
-    padding: 2rem 0 !important;
-    background-color: rgb(236, 98, 90) !important;
-  }
-
-  .card-body:hover {
-    background-color: rgb(242, 59, 59) !important;
-  }
-  
-  .card-text {
-    padding: 0.3rem 1rem;
-    font-size: 1rem;
-    color: black !important;
-  }
-
-  .overflow {
-    overflow: hidden;
-  }
-  
-  .card-img-top {
-    height: 100%;
-    transform: scale(1);
-    transition: all 0.2s ease-in-out;
-  }
-  
-  .card-img-top:hover {
-    transform: scale(1.8);
-  }
-
-  .col-md-4 {
-    padding: 0.5rem 0.5rem 0.5rem 0.5rem !important;
-  }
-
-  .btn {
-    border: none;
-    padding: 1rem 4rem;
-    font-size: 1.4rem;
-    color: white;
-    border-radius: 4rem;
-    transition: 0.5s ease-in-out;
-    cursor: pointer;
-    background: linear-gradient(to right, #fc4958, #e85d04);
-    text-transform: uppercase;
-  }
-
-  .btn:hover {
-    background: linear-gradient(to right, #e85d04, #fc4958);
-  }
