+++ rollingFood/src/components/card/card.css
@@ -1,0 +1,36 @@
+.img-card{
+    height: 100%;
+    object-fit: contain;
+    max-height: 20em;
+}
+.card-body-rolling{
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    justify-content: space-between;
+}
+.price{
+    font-size: 2em;
+}
+.card{
+    font-family:Arial, Helvetica, sans-serif ;
+    animation: backInLeft;
+    animation-duration: 2s; 
+}
+.title{
+    font-size: 3em;
+}
+.description{
+    font-size: 1.5em;
+}
+@media (max-width: 576px){
+    .title{
+        font-size: 1.5em;
+    }
+    .description{
+        font-size: 1em;
+    }
+    .price{
+        font-size: 1.3em;
+    }
+}
